Move get_task(s) to tasks file
